Add support for arrays in event_id
This is for cases where there is more than one event within a REDCap
project.  You must provide the Event Name and Event ID keys.  The Event
Name is found in the nested div of the <th> tag of the column.  The
Event_ID will be found on a record in the column.